Move layer to ol subdirectory
